+++ Classes/SchemaOrg/CreativeWork.php
@@ -9,32 +9,36 @@ namespace Portrino\PxSemantic\SchemaOrg;
  *
  * @author Andre Wuttig<wuttig@portrino.de>
  */
-class CreativeWork extends Thing {
-
+abstract class CreativeWork extends Thing
+{
     /**
      * @var AggregateRating The overall rating, based on a collection of reviews or ratings, of the item.
      */
     private $aggregateRating;
-
     /**
      * @var Person The author of this content. Please note that author is special in that HTML 5 provides a special mechanism for indicating authorship via the rel tag. That is equivalent to this and may be used interchangeably.
      */
     private $author;
-
     /**
      * @var \DateTime The date on which the CreativeWork was created or the item was added to a DataFeed.
      */
     private $dateCreated;
-
     /**
      * @var \DateTime The date on which the CreativeWork was most recently modified or when the item's entry was modified within a DataFeed.
      */
     private $dateModified;
-
     /**
      * @var \DateTime Date of first broadcast/publication.
      */
     private $datePublished;
+    /**
+     * @var string Headline of the article.
+     */
+    private $headline;
+    /**
+     * @var Organization The publisher of the creative work.
+     */
+    private $publisher;
 
     /**
      * Sets aggregateRating.
@@ -43,7 +47,8 @@ class CreativeWork extends Thing {
      *
      * @return $this
      */
-    public function setAggregateRating(AggregateRating $aggregateRating = NULL) {
+    public function setAggregateRating(AggregateRating $aggregateRating = null)
+    {
         $this->aggregateRating = $aggregateRating;
 
         return $this;
@@ -54,7 +59,8 @@ class CreativeWork extends Thing {
      *
      * @return AggregateRating
      */
-    public function getAggregateRating() {
+    public function getAggregateRating()
+    {
         return $this->aggregateRating;
     }
 
@@ -65,7 +71,8 @@ class CreativeWork extends Thing {
      *
      * @return $this
      */
-    public function setAuthor(Person $author = NULL) {
+    public function setAuthor(Person $author = null)
+    {
         $this->author = $author;
 
         return $this;
@@ -76,7 +83,8 @@ class CreativeWork extends Thing {
      *
      * @return Person
      */
-    public function getAuthor() {
+    public function getAuthor()
+    {
         return $this->author;
     }
 
@@ -87,7 +95,8 @@ class CreativeWork extends Thing {
      *
      * @return $this
      */
-    public function setDateCreated(\DateTime $dateCreated = NULL) {
+    public function setDateCreated(\DateTime $dateCreated = null)
+    {
         $this->dateCreated = $dateCreated;
 
         return $this;
@@ -98,7 +107,8 @@ class CreativeWork extends Thing {
      *
      * @return \DateTime
      */
-    public function getDateCreated() {
+    public function getDateCreated()
+    {
         return $this->dateCreated;
     }
 
@@ -109,7 +119,8 @@ class CreativeWork extends Thing {
      *
      * @return $this
      */
-    public function setDateModified(\DateTime $dateModified = NULL) {
+    public function setDateModified(\DateTime $dateModified = null)
+    {
         $this->dateModified = $dateModified;
 
         return $this;
@@ -120,7 +131,8 @@ class CreativeWork extends Thing {
      *
      * @return \DateTime
      */
-    public function getDateModified() {
+    public function getDateModified()
+    {
         return $this->dateModified;
     }
 
@@ -131,7 +143,8 @@ class CreativeWork extends Thing {
      *
      * @return $this
      */
-    public function setDatePublished(\DateTime $datePublished = NULL) {
+    public function setDatePublished(\DateTime $datePublished = null)
+    {
         $this->datePublished = $datePublished;
 
         return $this;
@@ -142,7 +155,56 @@ class CreativeWork extends Thing {
      *
      * @return \DateTime
      */
-    public function getDatePublished() {
+    public function getDatePublished()
+    {
         return $this->datePublished;
+    }
+
+    /**
+     * Sets headline.
+     *
+     * @param string $headline
+     *
+     * @return $this
+     */
+    public function setHeadline($headline)
+    {
+        $this->headline = $headline;
+
+        return $this;
+    }
+
+    /**
+     * Gets headline.
+     *
+     * @return string
+     */
+    public function getHeadline()
+    {
+        return $this->headline;
+    }
+
+    /**
+     * Sets publisher.
+     *
+     * @param Organization $publisher
+     *
+     * @return $this
+     */
+    public function setPublisher(Organization $publisher = null)
+    {
+        $this->publisher = $publisher;
+
+        return $this;
+    }
+
+    /**
+     * Gets publisher.
+     *
+     * @return Organization
+     */
+    public function getPublisher()
+    {
+        return $this->publisher;
     }
 }