+++ Classes/SchemaOrg/Thing.php
@@ -15,44 +15,36 @@ abstract class Thing extends \TYPO3\CMS\Extbase\DomainObject\AbstractEntity {
      * @var string An additional type for the item, typically used for adding more specific types from external vocabularies in microdata syntax. This is a relationship between something and a class that the thing is in. In RDFa syntax, it is better to use the native RDFa syntax - the 'typeof' attribute - for multiple types. Schema.org tools may have only weaker understanding of extra types, in particular those defined externally.
      */
     private $additionalType;
-
     /**
      * @var string An alias for the item.
      */
     private $alternateName;
-
     /**
      * @var string A short description of the item.
      */
     private $description;
-
     /**
      * @var ImageObject An image of the item. This can be a [URL](http://schema.org/URL) or a fully described [ImageObject](http://schema.org/ImageObject).
      */
     private $image;
-
     /**
-     * @var string Indicates a page (or other CreativeWork) for which this thing is the main entity being described.
+     * @var WebPage Indicates a page (or other CreativeWork) for which this thing is the main entity being described.
      *
-     *             See [background notes](/docs/datamodel.html#mainEntityBackground) for details.
+     *              See [background notes](/docs/datamodel.html#mainEntityBackground) for details.
      */
     private $mainEntityOfPage;
-
     /**
      * @var string The name of the item.
      */
     private $name;
-
     /**
-     * @var string|array multiple URLs of a reference Web page that unambiguously indicates the item's identity. E.g. the URL of the item's Wikipedia page, Freebase page, or official website.
+     * @var string URL of a reference Web page that unambiguously indicates the item's identity. E.g. the URL of the item's Wikipedia page, Freebase page, or official website.
      */
     private $sameAs;
-
     /**
      * @var string URL of the item.
      */
     private $url;
-
     /**
      * @var Action Indicates a potential Action, which describes an idealized action in which this thing would play an 'object' role.
      */
@@ -149,11 +141,11 @@ abstract class Thing extends \TYPO3\CMS\Extbase\DomainObject\AbstractEntity {
     /**
      * Sets mainEntityOfPage.
      *
-     * @param string $mainEntityOfPage
+     * @param WebPage $mainEntityOfPage
      *
      * @return $this
      */
-    public function setMainEntityOfPage($mainEntityOfPage) {
+    public function setMainEntityOfPage(WebPage $mainEntityOfPage = NULL) {
         $this->mainEntityOfPage = $mainEntityOfPage;
 
         return $this;
@@ -162,7 +154,7 @@ abstract class Thing extends \TYPO3\CMS\Extbase\DomainObject\AbstractEntity {
     /**
      * Gets mainEntityOfPage.
      *
-     * @return string
+     * @return WebPage
      */
     public function getMainEntityOfPage() {
         return $this->mainEntityOfPage;
@@ -193,7 +185,7 @@ abstract class Thing extends \TYPO3\CMS\Extbase\DomainObject\AbstractEntity {
     /**
      * Sets sameAs.
      *
-     * @param string|array $sameAs
+     * @param string $sameAs
      *
      * @return $this
      */
@@ -206,7 +198,7 @@ abstract class Thing extends \TYPO3\CMS\Extbase\DomainObject\AbstractEntity {
     /**
      * Gets sameAs.
      *
-     * @return string|array
+     * @return string
      */
     public function getSameAs() {
         return $this->sameAs;