+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # PxSemantic Change log
 
+1.1.0 - 2016-02-05
+-------------------
+* adds feature to set an id for an entity and the json+ld encoder will therefore set the "@id" property which was defined by json-ld standard
+* adds Article, TechArticle, WebPage, Website entities
+* adds some properties to Thing, CreativeWork
+* fixes wrong inheritance for DataType
+
 1.0.0 - 2016-01-22
 -------------------
 * initial import